chore: [vgf-runtime] prepare shared test utilities
Extract common Vulkan/SPIR-V test helpers and VGF encoder helpers before introducing the new runtime API.

This keeps the old VGF runtime tests behavior-preserving while reducing helper churn in the foundation patch.
